Fixed ItemDPromotion calculation

+++ CheckoutKata/Services/Abstractions/ICheckout.cs
@@ -6,7 +6,7 @@ namespace CheckoutKata.Services.Abstractions
     public interface ICheckout
     {
         void AddToCheckout(string item);
-        int GetTotal();
+        double GetTotal();
         List<CheckoutItem> GetCheckout();
     }
 }
